Phase_3/requirements.txt
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
WTForms==3.0.1
Werkzeug==2.0.0
mysql-connector-c==6.1.11
mysql-connector-python==8.0.1
numpy==1.20.3
pandas==1.3.4
datetime==4.4
